refactor(direct-control): extract setup prepare owner
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only prepareCiv7SinglePlayerSetup, its generated setup preparation command, setup input normalization, setup map-row/readback assertions, and private setup parameter helpers into packages/civ7-direct-control/src/setup/prepare.ts while keeping the public facade export in index.ts. startPreparedCiv7SinglePlayerGame and runCiv7SinglePlayerFromSetup still reuse the internal normalization/readback helpers until their owner slices move.

Preserves approval-first setup mutation, setup snapshot readback, setup map-row proof, setup option validation, command string/source shape, and no-replay-after-socket-close package proof. Start/run, restart/begin lifecycle orchestration, remaining no-replay ownership, hotseat runtime proof, AI ingestion, semantic CLI projection, telemetry, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- setup-and-lifecycle.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
- ownership and no-cycle rg scans

Local package/source relocation proof only; no runtime/live-game, lifecycle behavior-change, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
