+++ packages/civ7-direct-control/src/index.ts
@@ -10,11 +10,6 @@ import {
 } from "./civ7-component-id.js";
 import { Civ7DirectControlError, type Civ7DirectControlErrorCode } from "./direct-control-error.js";
 import {
-  CIV7_SIGNED_INT_SEED_MAX,
-  CIV7_SIGNED_INT_SEED_MIN,
-  assessCiv7SignedIntSeed,
-} from "./policy/setup.js";
-import {
   diplomacyResponsePostcondition,
   waitForCiv7DiplomacyResponseAfter,
 } from "./play/operations/diplomacy-postconditions.js";
@@ -61,9 +56,13 @@ import {
   ensureCiv7SetupMapRowVisible as ensureCiv7SetupMapRowVisibleFromModule,
   getCiv7SetupMapRows as getCiv7SetupMapRowsFromModule,
   getCiv7SetupSnapshot as getCiv7SetupSnapshotFromModule,
-  setupSnapshotScriptSource,
-  validateMapScript,
 } from "./setup/reads.js";
+import {
+  assertPreparedSetupMatches,
+  normalizeSavedGameConfigurationRef,
+  normalizeSinglePlayerSetupInput as normalizeSinglePlayerSetupInputFromModule,
+  prepareCiv7SinglePlayerSetup as prepareCiv7SinglePlayerSetupFromModule,
+} from "./setup/prepare.js";
 import {
   configureCiv7Autoplay as configureCiv7AutoplayFromModule,
   getCiv7AutoplayStatus as getCiv7AutoplayStatusFromModule,
@@ -2996,14 +2995,22 @@ function setupReadDependencies() {
     executeAppUiCommand: executeCiv7AppUiCommand,
     exitToMainMenuCommand: CIV7_EXIT_TO_MAIN_MENU_COMMAND,
     jsLiteral,
+    loadSavedGameConfiguration: loadCiv7SavedGameConfiguration,
     parseSetupMapRows: (result: Civ7CommandResult, label: string) =>
       jsonPayloadFromCommandResult<Civ7SetupMapRowsResult>(result, label),
+    parseSetupPreparation: (result: Civ7CommandResult, label: string) =>
+      jsonPayloadFromCommandResult<{
+        before: Civ7SetupSnapshot;
+        after: Civ7SetupSnapshot;
+        applied: Record<string, Civ7SetupOptionValue>;
+      }>(result, label),
     parseSetupSnapshot: (result: Civ7CommandResult, label: string) =>
       jsonPayloadFromCommandResult<Civ7SetupSnapshotResult>(result, label),
     probeHelperSource,
     playerSetupParameterIds: DEFAULT_CIV7_PLAYER_SETUP_PARAMETER_IDS,
     reloadUiCommand: CIV7_RELOAD_UI_COMMAND,
     setupParameterIds: DEFAULT_CIV7_SETUP_PARAMETER_IDS,
+    validateIdentifier,
   } as const;
 }
 
@@ -3069,66 +3076,7 @@ export async function prepareCiv7SinglePlayerSetup(
   options: Civ7DirectControlOptions = {},
   approval: Civ7ActionApproval,
 ): Promise<Civ7PreparedSetupResult> {
-  assertApproved(approval, "preparing a Civ7 single-player setup");
-  const normalized = normalizeSinglePlayerSetupInput(input);
-  const savedConfigLoad = normalized.savedConfig
-    ? await loadCiv7SavedGameConfiguration(normalized.savedConfig, options, {
-        waitTimeoutMs: options.timeoutMs,
-        pollIntervalMs: 1_000,
-      })
-    : undefined;
-  if (savedConfigLoad && !savedConfigLoad.loaded) {
-    throw new Civ7DirectControlError(
-      "setup-config-load-failed",
-      `Civ7 did not load saved configuration ${savedConfigLoad.savedConfig.fileName}`,
-      { details: savedConfigLoad },
-    );
-  }
-  const before = await getCiv7SetupSnapshot(options);
-  if (normalized.requireShell !== false && before.snapshot.phase !== "shell") {
-    throw new Civ7DirectControlError(
-      "setup-phase-invalid",
-      `Civ7 setup requires shell/main-menu phase; observed ${before.snapshot.phase}`,
-      { details: before },
-    );
-  }
-
-  const rowProof = findSetupMapRow(before.snapshot, normalized.mapScript);
-  if (!rowProof) {
-    throw new Civ7DirectControlError(
-      "setup-map-row-missing",
-      `Civ7 setup map row is not visible for ${normalized.mapScript}`,
-      { details: before.snapshot.mapRows },
-    );
-  }
-
-  const command = await executeCiv7AppUiCommand({
-    ...options,
-    command: buildPrepareSinglePlayerSetupCommand(normalized),
-  });
-  const payload = jsonPayloadFromCommandResult<{
-    before: Civ7SetupSnapshot;
-    after: Civ7SetupSnapshot;
-    applied: Record<string, Civ7SetupOptionValue>;
-  }>(command, "Civ7 setup preparation");
-  const after = {
-    host: command.host,
-    port: command.port,
-    state: command.state,
-    snapshot: payload.after,
-  };
-  assertPreparedSetupMatches(normalized, after.snapshot);
-  return {
-    host: command.host,
-    port: command.port,
-    state: command.state,
-    before,
-    after,
-    command,
-    ...(savedConfigLoad ? { savedConfigLoad } : {}),
-    applied: payload.applied,
-    verified: true,
-  };
+  return await prepareCiv7SinglePlayerSetupFromModule(input, options, approval, setupReadDependencies());
 }
 
 export async function startPreparedCiv7SinglePlayerGame(
@@ -3137,7 +3085,7 @@ export async function startPreparedCiv7SinglePlayerGame(
   approval: Civ7ActionApproval,
 ): Promise<Civ7SinglePlayerStartResult> {
   assertApproved(approval, "starting a prepared Civ7 single-player game");
-  const expected = normalizeSinglePlayerSetupInput(input.expected);
+  const expected = normalizeSinglePlayerSetupInputFromModule(input.expected, setupReadDependencies());
   const before = await getCiv7SetupSnapshot(options);
   assertPreparedSetupMatches(expected, before.snapshot);
 
@@ -3241,7 +3189,7 @@ export async function runCiv7SinglePlayerFromSetup(
   approval: Civ7ActionApproval,
 ): Promise<Civ7SinglePlayerRunResult> {
   assertApproved(approval, "running Civ7 single-player from setup");
-  const normalized = normalizeSinglePlayerSetupInput(input);
+  const normalized = normalizeSinglePlayerSetupInputFromModule(input, setupReadDependencies());
   let shellExit: Civ7CommandResult | undefined;
   const initial = await getCiv7SetupSnapshot(options);
   if (initial.snapshot.phase !== "shell") {
@@ -4181,59 +4129,6 @@ function buildLoadSavedGameConfigurationCommand(input: Civ7SavedGameConfiguratio
     });
   })()`;
 }
-
-function buildPrepareSinglePlayerSetupCommand(input: Civ7SinglePlayerSetupInput): string {
-  return `(() => {
-    ${setupSnapshotScriptSource(setupReadDependencies())}
-    const input = ${jsLiteral(input)};
-    const before = readSetupSnapshot();
-    const applied = {};
-    const setSetupParameter = (id, value) => {
-      if (typeof GameSetup !== "undefined" && GameSetup && typeof GameSetup.setGameParameterValue === "function") {
-        GameSetup.setGameParameterValue(id, value);
-      }
-    };
-    const setPlayerSetupParameter = (playerId, id, value) => {
-      if (typeof GameSetup !== "undefined" && GameSetup && typeof GameSetup.setPlayerParameterValue === "function") {
-        GameSetup.setPlayerParameterValue(playerId, id, value);
-      }
-    };
-    const editMap = Configuration.editMap();
-    const editGame = Configuration.editGame();
-    if (!editMap || !editGame) throw new Error("Configuration edit APIs are unavailable");
-    editMap.setScript(input.mapScript);
-    setSetupParameter("Map", input.mapScript);
-    applied.Map = input.mapScript;
-    editMap.setMapSize(input.mapSize);
-    setSetupParameter("MapSize", input.mapSize);
-    applied.MapSize = input.mapSize;
-    editMap.setMapSeed(input.seed);
-    setSetupParameter("MapRandomSeed", input.seed);
-    applied.MapRandomSeed = input.seed;
-    if (input.gameSeed !== undefined) {
-      editGame.setGameSeed(input.gameSeed);
-      setSetupParameter("GameRandomSeed", input.gameSeed);
-      applied.GameRandomSeed = input.gameSeed;
-    }
-    if (input.playerCount !== undefined && typeof editMap.setMaxMajorPlayers === "function") {
-      editMap.setMaxMajorPlayers(input.playerCount);
-      applied.MaxMajorPlayers = input.playerCount;
-    }
-    for (const [key, value] of Object.entries(input.options ?? {})) {
-      GameSetup.setGameParameterValue(key, value);
-      applied[key] = value;
-    }
-    for (const player of input.playerOptions ?? []) {
-      for (const [key, value] of Object.entries(player.options ?? {})) {
-        setPlayerSetupParameter(player.playerId, key, value);
-        applied["Player:" + player.playerId + ":" + key] = value;
-      }
-    }
-    const after = readSetupSnapshot();
-    return JSON.stringify({ before, after, applied });
-  })()`;
-}
-
 function buildStartPreparedSinglePlayerCommand(): string {
   return `(() => {
     const serverType = typeof ServerType !== "undefined" && ServerType && ServerType.SERVER_TYPE_NONE !== undefined
@@ -4894,170 +4789,6 @@ function summarizeCiv7CfgStrings(strings: ReadonlyArray<string>): Civ7SavedGameC
   if (gameSeed !== undefined) summary.gameSeed = gameSeed;
   return summary;
 }
-
-function validateSavedConfigFileName(value: string): string {
-  if (
-    !value.trim() ||
-    value.length > 512 ||
-    /[\0\r\n]/.test(value) ||
-    value.includes("/") ||
-    value.includes("\\") ||
-    extname(value).toLowerCase() !== ".civ7cfg"
-  ) {
-    throw new Civ7DirectControlError("setup-parameter-invalid", "saved configuration fileName must be a Civ7Cfg file name");
-  }
-  return value;
-}
-
-function normalizeSavedGameConfigurationRef(input: Civ7SavedGameConfigurationRef): Civ7SavedGameConfigurationRef {
-  const fileName = validateSavedConfigFileName(input.fileName);
-  const displayName = input.displayName.trim() || basename(fileName, extname(fileName));
-  const path = input.path.trim();
-  if (!path || /[\0\r\n]/.test(path)) {
-    throw new Civ7DirectControlError("setup-parameter-invalid", "saved configuration path must be a non-empty single-line string");
-  }
-  return {
-    id: input.id.trim() || basename(fileName, extname(fileName)),
-    displayName,
-    fileName,
-    path,
-  };
-}
-
-function normalizeSinglePlayerSetupInput(input: Civ7SinglePlayerSetupInput): Civ7SinglePlayerSetupInput {
-  validateMapScript(input.mapScript);
-  if (!/^MAPSIZE_[A-Z0-9_]+$/.test(input.mapSize)) {
-    throw new Civ7DirectControlError("setup-parameter-invalid", "mapSize must be a Civ7 MAPSIZE_* value");
-  }
-  validateSetupSeed(input.seed, "seed");
-  if (input.gameSeed !== undefined) validateSetupSeed(input.gameSeed, "gameSeed");
-  if (input.playerCount !== undefined) boundedInteger(input.playerCount, 1, 64, "playerCount");
-  const options: Record<string, Civ7SetupOptionValue> = {};
-  for (const [key, value] of Object.entries(input.options ?? {})) {
-    validateIdentifier(key, "setup option id");
-    if (!["string", "number", "boolean"].includes(typeof value)) {
-      throw new Civ7DirectControlError("setup-parameter-invalid", `Unsupported setup option value for ${key}`);
-    }
-    options[key] = value;
-  }
-  const playerOptions: Civ7PlayerSetupOptions[] = [];
-  for (const player of input.playerOptions ?? []) {
-    const playerId = boundedInteger(player.playerId, 0, 64, "playerOptions.playerId");
-    const normalizedOptions: Record<string, Civ7SetupOptionValue> = {};
-    for (const [key, value] of Object.entries(player.options ?? {})) {
-      validateIdentifier(key, "player setup option id");
-      if (!["string", "number", "boolean"].includes(typeof value)) {
-        throw new Civ7DirectControlError("setup-parameter-invalid", `Unsupported player setup option value for ${key}`);
-      }
-      normalizedOptions[key] = value;
-    }
-    if (Object.keys(normalizedOptions).length > 0) {
-      playerOptions.push({ playerId, options: normalizedOptions });
-    }
-  }
-  return {
-    ...input,
-    mapScript: input.mapScript,
-    mapSize: input.mapSize,
-    seed: input.seed,
-    ...(input.savedConfig ? { savedConfig: normalizeSavedGameConfigurationRef(input.savedConfig) } : {}),
-    options,
-    playerOptions,
-  };
-}
-
-function validateSetupSeed(value: unknown, label: string): number {
-  const seed = assessCiv7SignedIntSeed(value);
-  if (seed.ok === false) {
-    const suffix = seed.reason === "not-integer"
-      ? "must be an integer"
-      : `must be an integer between ${seed.min} and ${seed.max}`;
-    throw new Civ7DirectControlError("setup-parameter-invalid", `${label} ${suffix}`);
-  }
-  return seed.value;
-}
-
-function findSetupParameter(snapshot: Civ7SetupSnapshot, id: string): Civ7SetupParameterSnapshot | undefined {
-  return snapshot.setup.parameters.find((parameter) => parameter.id === id);
-}
-
-function setupParameterValue(snapshot: Civ7SetupSnapshot, id: string): Civ7SetupParameterValue | undefined {
-  return findSetupParameter(snapshot, id)?.value;
-}
-
-function findPlayerSetupParameter(snapshot: Civ7SetupSnapshot, playerId: number, id: string): Civ7SetupParameterSnapshot | undefined {
-  return snapshot.setup.playerParameters
-    .find((player) => player.playerId === playerId)
-    ?.parameters.find((parameter) => parameter.id === id);
-}
-
-function playerSetupParameterValue(
-  snapshot: Civ7SetupSnapshot,
-  playerId: number,
-  id: string,
-): Civ7SetupParameterValue | undefined {
-  return findPlayerSetupParameter(snapshot, playerId, id)?.value;
-}
-
-function findSetupMapRow(snapshot: Civ7SetupSnapshot, file: string): Civ7SetupMapRow | undefined {
-  return snapshot.mapRows.find((row) => row.file === file || row.value === file);
-}
-
-function assertPreparedSetupMatches(input: Civ7SinglePlayerSetupInput, snapshot: Civ7SetupSnapshot): void {
-  const mapRow = findSetupMapRow(snapshot, input.mapScript);
-  if (!mapRow) {
-    throw new Civ7DirectControlError(
-      "setup-map-row-missing",
-      `Civ7 setup map row did not read back for ${input.mapScript}`,
-      { details: snapshot.mapRows },
-    );
-  }
-  const script = setupParameterValue(snapshot, "Map");
-  const mapSize = setupParameterValue(snapshot, "MapSize");
-  const mapSeed = setupParameterValue(snapshot, "MapRandomSeed");
-  const gameSeed = setupParameterValue(snapshot, "GameRandomSeed");
-  if (script !== input.mapScript) {
-    throw new Civ7DirectControlError("setup-readback-mismatch", `Civ7 setup Map readback mismatch: ${String(script)}`, {
-      details: { expected: input.mapScript, actual: script, snapshot },
-    });
-  }
-  if (mapSize !== input.mapSize) {
-    throw new Civ7DirectControlError("setup-readback-mismatch", `Civ7 setup MapSize readback mismatch: ${String(mapSize)}`, {
-      details: { expected: input.mapSize, actual: mapSize, snapshot },
-    });
-  }
-  if (Number(mapSeed) !== input.seed) {
-    throw new Civ7DirectControlError("setup-readback-mismatch", `Civ7 setup MapRandomSeed readback mismatch: ${String(mapSeed)}`, {
-      details: { expected: input.seed, actual: mapSeed, snapshot },
-    });
-  }
-  if (input.gameSeed !== undefined && Number(gameSeed) !== input.gameSeed) {
-    throw new Civ7DirectControlError("setup-readback-mismatch", `Civ7 setup GameRandomSeed readback mismatch: ${String(gameSeed)}`, {
-      details: { expected: input.gameSeed, actual: gameSeed, snapshot },
-    });
-  }
-  for (const [key, expected] of Object.entries(input.options ?? {})) {
-    const actual = setupParameterValue(snapshot, key);
-    if (actual !== expected) {
-      throw new Civ7DirectControlError("setup-readback-mismatch", `Civ7 setup ${key} readback mismatch: ${String(actual)}`, {
-        details: { expected, actual, snapshot },
-      });
-    }
-  }
-  for (const player of input.playerOptions ?? []) {
-    for (const [key, expected] of Object.entries(player.options)) {
-      const actual = playerSetupParameterValue(snapshot, player.playerId, key);
-      if (actual !== expected) {
-        throw new Civ7DirectControlError(
-          "setup-readback-mismatch",
-          `Civ7 player ${player.playerId} setup ${key} readback mismatch: ${String(actual)}`,
-          { details: { playerId: player.playerId, expected, actual, snapshot } },
-        );
-      }
-    }
-  }
-}
-
 function assertPostStartMatches(input: Civ7SinglePlayerSetupInput, summary: Civ7MapSummaryResult): void {
   const seed = probeValue(summary.map.randomSeed);
   if (seed !== undefined && seed !== input.seed) {
